remove fit from time series classification
