+++ evalml/pipelines/time_series_classification_pipelines.py
@@ -31,37 +31,6 @@ class TimeSeriesClassificationPipeline(TimeSeriesPipelineBase, ClassificationPip
         random_seed (int): Seed for the random number generator. Defaults to 0.
     """
 
-    def fit(self, X, y):
-        """Fit a time series classification pipeline.
-
-        Args:
-            X (pd.DataFrame or np.ndarray): The input training data of shape [n_samples, n_features].
-            y (pd.Series, np.ndarray): The target training targets of length [n_samples].
-
-        Returns:
-            self
-
-        Raises:
-            ValueError: If the number of unique classes in y are not appropriate for the type of pipeline.
-        """
-        X, y = self._convert_to_woodwork(X, y)
-
-        if self.problem_type == ProblemTypes.TIME_SERIES_BINARY and y.nunique() != 2:
-            raise ValueError(
-                f"Time Series Binary pipelines require y to have 2 unique classes!"
-            )
-        elif (
-            self.problem_type == ProblemTypes.TIME_SERIES_MULTICLASS
-            and y.nunique() in [1, 2]
-        ):
-            raise ValueError(
-                "Time Series Multiclass pipelines require y to have 3 or more unique classes!"
-            )
-
-        self._fit(X, y)
-        self._classes_ = list(ww.init_series(np.unique(y)))
-        return self
-
     def _estimator_predict_proba(self, features):
         """Get estimator predicted probabilities.
